Fixed another bug about newlines
I have no idea but for some reason sometimes there are excesive newlines, and that is being added bt WrtitePart.

The fact is since we aren't using the regex match for the last part, the newlines now get passed through. Due to because some lines do end with newline and others do not, I made a check to see if they end with a newline, just output w/o adding newline and if it is lacking one, add one!

This bug was put in evidence by the 1.4.4 bug fix

Note: For some reason I had to re-pull to notice this bug, the files were 100% the same and it worked fine last time. I guess computers can be wierd.